Secure traffic to the Splunk monitor page
As mentioned elsewhere, the master is attached to the monitor ELB.  The master
node provides sufficient information about the cluster and the peers' health.
If we need to in the future, We could add a new Splunk monitor instance and
attach it to the ELB instead of the master node.

[finishes #167171028](https://www.pivotaltracker.com/story/show/167171028)

Authored-by: John Shahid <jshahid@pivotal.io>