Set allowInternetAccess to false to prevent Splunk from using the internet
I could only attempts to connect to apps.splunk.com from the indexer and search
head VMs.  In addition to the indexers and search head, I added the
configuration to the master and forwarder for consistency.

Also turned on compression and base64 encoding of the user data in all four
modules.  The Splunk master user data was getting bigger.  We originally
disabled compression because the user data was later wrapped in another
multipart cloud config in the launch module, but that isn't the case anymore.

[#169126797](https://www.pivotaltracker.com/story/show/169126797)

Authored-by: John Shahid <jshahid@pivotal.io>