Setup a EC2 VPC endpoint inside the control plane VPC
So that terragrunt can still reach EC2 when recreating the NAT instances.
Otherwise, terragrunt will just hang while running from inside the CP VPC.  For
example when terragrunt runs on the SJB or Concourse.

Since we only create a VPC endpoint for the EC2 service, terragrunt could fail
when running other layers, e.g., because it failed to reach STS.  In those
cases, terragrunt should leave the terraform state in a healthy condition
allowing the operator or the pipeline to attempt running terragrunt again.

[#169394505](https://www.pivotaltracker.com/story/show/169394505)

Authored-by: John Shahid <jshahid@pivotal.io>